backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.4
psycopg2-binary==2.9.10
alembic==1.14.0
python-multipart==0.0.12
requests==2.32.3
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
email-validator==2.1.1
python-dotenv==1.0.1
openai==1.51.2
